Implement authorization using Gates

+++ app/Http/Controllers/QuestionController.php
@@ -6,6 +6,7 @@ use App\Question;
 use Illuminate\Http\Request;
 use App\Http\Requests\AskQuestionRequest;
 
+
 class QuestionController extends Controller
 {
     /**
@@ -67,7 +68,13 @@ class QuestionController extends Controller
      */
     public function edit(Question $question)
     {
+
+        if(\Gate::denies('update-question', $question)) {     
+            abort(403, "Access denied!");
+        }
+
         return view('questions.edit', compact('question'));
+
     }
 
     /**
@@ -79,6 +86,10 @@ class QuestionController extends Controller
      */
     public function update(AskQuestionRequest $request, Question $question)
     {
+        if(\Gate::denies('update-question', $question)) {     
+            abort(403, "Access denied!");
+        }
+
         $question->update($request->only('title', 'body'));
 
         return redirect('/questions')->with('success', 'Your question has been updated');
@@ -92,6 +103,10 @@ class QuestionController extends Controller
      */
     public function destroy(Question $question)
     {
+        if(\Gate::denies('delete-question', $question)) {     
+            abort(403, "Access denied!");
+        }
+
         $question->delete();
 
         return redirect()->route('questions.index')->with('success', "Your question has been deleted!");
